fix spec problems
