+++ updater/spec/dependabot/updater/operations/create_security_update_pull_request_spec.rb
@@ -15,40 +15,6 @@ require "dependabot/notices"
 
 require "dependabot/bundler"
 
-# Stub PackageManagerBase
-class StubPackageManager < Dependabot::PackageManagerBase
-  def initialize(name:, version:, deprecated_versions: [], unsupported_versions: [], supported_versions: [])
-    @name = name
-    @version = version
-    @deprecated_versions = deprecated_versions
-    @unsupported_versions = unsupported_versions
-    @supported_versions = supported_versions
-  end
-
-  attr_reader :name
-  attr_reader :version
-  attr_reader :deprecated_versions
-  attr_reader :unsupported_versions
-  attr_reader :supported_versions
-
-  sig { override.returns(T::Boolean) }
-  def deprecated?
-    # If the version is unsupported, the unsupported error is getting raised separately.
-    return false if unsupported?
-
-    deprecated_versions.include?(version)
-  end
-
-  sig { override.returns(T::Boolean) }
-  def unsupported?
-    # Check if the feature flag for Bundler v1 unsupported error is enabled.
-    return false unless Dependabot::Experiments.enabled?(:bundler_v1_unsupported_error)
-
-    # Determine if the Bundler version is unsupported.
-    version < supported_versions.first
-  end
-end
-
 RSpec.describe Dependabot::Updater::Operations::CreateSecurityUpdatePullRequest do
   include DependencyFileHelpers
   include DummyPkgHelpers
@@ -95,7 +61,7 @@ RSpec.describe Dependabot::Updater::Operations::CreateSecurityUpdatePullRequest 
   end
 
   let(:package_manager) do
-    StubPackageManager.new(
+    DummyPkgHelpers::StubPackageManager.new(
       name: "bundler",
       version: package_manager_version,
       deprecated_versions: deprecated_versions,
